+++ db/migrate/20220730092904_devise_create_users.rb
@@ -33,7 +33,7 @@ class DeviseCreateUsers < ActiveRecord::Migration[7.0]
       # t.datetime :locked_at
 
       t.integer :role, default: 0
-      t.integer :balance, default: 0
+      t.decimal :balance, default: 0
       t.timestamps null: false
     end
